# api.yml
openapi: 3.0.2
info:
  description: Simple instant messaging application
  version: 2.0.0
  title: Simply Chat
  contact:
    email: biral.mattia@gmail.com
  license:
    name: GPL-3.0
    url: https://www.gnu.org/licenses/gpl-3.0.en.html
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: simplychat-auth
  schemas:
    username:
      type: string
      example: Pyrix25633
      minLength: 3
      maxLength: 32
    email:
      type: string
      format: email
      example: name.surname@domain.com
    password:
      type: string
      example: passwd1234@
      minLength: 8
...
paths:
...
  /temp-users:
    post:
      tags:
        - temp-users
      summary: Creates a new temporary user
      description: First step of the registration process
      operationId: temp-users
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                username:
                  $ref: "#/components/schemas/username"
                email:
                  $ref: "#/components/schemas/email"
                password:
                  $ref: "#/components/schemas/password"
              required:
                - username
                - email
                - password
      responses:
        "201":
          description: The first step of the registration succeded
          content:
            application/json:
              schema:
                type: object
                properties:
                  username:
                    $ref: "#/components/schemas/username"
        "400":
          description: Bad Request
        "422":
          description: "Unprocessable Content: temporary user cannot be created because of duplicate username or email"
  /temp-users/{username}/confirm:
    post:
      tags:
        - temp-users
      summary: Confirms a temporary user
      description: Sencond and last step of the registration process
      operationId: temp-users/[username]/confirm
      parameters:
        - in: path
          name: username
          schema:
            $ref: "#/components/schemas/username"
          required: true
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                verificationCode:
                  type: integer
                  minimum: 100000
                  maximum: 999999
                  example: 190347
              required:
                - verificationCode
      responses:
        "201":
          description: The second step of the registration succeded, user created
          content:
            application/json:
              schema:
                type: object
                properties:
                  userId:
                    type: integer
                    example: 10
        "400":
          description: Bad Request
        "404":
          description: "Not Found: no temporary users found with specified username"
        "422":
          description: "Unprocessable Content: user cannot be created"
...
servers:
  - url: https://simplychat.ddns.net/api/